use longs for most things instead of integers

+++ fixture/MessageTemplateParameterSpec.java
@@ -24,6 +24,6 @@ class MessageTemplateParameterSpec {
 	String description;
 
 	@DataAttribute
-	Integer maximumLength;
+	Long maximumLength;
 
 }
